packages/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pycryptodome>=3.20",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]
